Arduino Based Smart Lab Security System
The very important component of this system is ARDIUNO UNO. It is used as the
controller in the security system. The all modules and components relate to the ARDIUNO UNO
development board. The PIR sensors used to sense the movement of any moving object to
controlling the light in lab or certain room. If the PIR sensor sense any moving object it will give
signal to relay through the ARDIUNO UNO. Here the use of relay to isolate the 5v signal to 110v
to 230v supply for controlling the light. The IR counter is used to measure the persons entered in
the lab. The count is displayed on the 16x2 liquid crystal display. The display is also connected to
the ARDIUNO UNO board.
 The fingerprint sensor named as R-305 fingerprint sensor is used to take fingers
impressions to verifying the finger prints which previously stored in the database of ARDIUNO
UNO. If the finger impression is matched with the database impression the signal will be generated
and go to the ARDIUNO UNO board. The controller will give the command to the solenoid lock.
If the command positive means the impression is matched the lock will open. And, if the command
is negative in nature then the lock didn’t change its status. Here the solenoid lock is used to lock
the doors. This sensor is also used for securing the documents cupboards.